% Generated by build script
\name{hmda_aggregations}
\alias{hmda_aggregations}
\title{Fetch HMDA nationwide aggregations}
\usage{
hmda_aggregations()
}
\arguments{
\item{years}{Integer year (e.g. 2022). Required.}
\item{actions_taken}{Action taken code: "1" (originated), "2" (approved),}
\item{loan_types}{Loan type: "1" (conventional), "2" (FHA), "3" (VA), "4" (USDA)}
\item{purposes}{Purpose: "1" (purchase), "2" (improvement), "31" (refinance),}
\item{races}{Race code: "1" (Native), "2" (Asian), "3" (Black),}
\item{ethnicities}{Ethnicity: "1" (Hispanic), "2" (Not Hispanic),}
}
\value{tibble: count (integer), sum (numeric), plus filter columns}
\description{Fetch HMDA nationwide aggregations}
