% Generated by build script
\name{hmda_filers}
\alias{hmda_filers}
\title{Fetch HMDA filer institutions for a given year}
\usage{
hmda_filers()
}
\arguments{
\item{years}{Integer year (e.g. 2022)}
}
\value{tibble: lei (character), name (character), period (integer)}
\description{Fetch HMDA filer institutions for a given year}
